backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.2
Flask-SocketIO==5.4.1
Flask-JWT-Extended==4.6.0
python-ldap==3.4.3
psycopg2-binary==2.9.9
PyJWT==2.11.0
bcrypt==5.0.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-socketio==5.10.0
python-engineio==4.8.0
email-validator==2.1.0
requests==2.31.0
PyPDF2==4.0.1
openpyxl==3.1.2
python-dateutil==2.8.2
pydantic==2.5.3
